requirements.txt
#asgiref==3.2.7
#astroid==2.3.3
#atomicwrites==1.3.0
#attrs==19.3.0
#boto3==1.12.36
boto3
#botocore==1.15.36
#cachetools#==4.0.0
certifi
chardet
colorama
Django
django-cors-headers
django-storages
djangorestframework
docutils
entrypoints
#facebook-sdk==3.1.0
#flake8==3.7.9
google-auth
idna
importlib-metadata
isort
jmespath
# jsonschema==2.6.0
jsonschema

#lazy-object-proxy==1.4.3
#mccabe==0.6.1
#more-itertools==8.2.0
packaging
pluggy
#psycopg2==2.8.4
py
pyasn1
pyasn1-modules
pycodestyle
pyflakes
pylint
pyparsing
#pytest==5.2.1
python-dateutil
python-http-client
pytz
PyYAML
requests
rsa
#s3transfer==0.3.3
#sendgrid==6.0.5
#sentry-sdk==0.14.2
#six==1.14.0
sqlparse
#typed-ast==1.4.1
urllib3
#wcwidth==0.1.9
#wrapt==1.11.2
#zipp==3.1.0
django_extensions
# pycryptodome
django-migrate-sql-deux
#wiki
html-testRunner
#--pre wiki
redgreenunittest
pyjwt
drf-yasg
#django_cognito_jwt
xhtml2pdf
django-cors-headers